requirements (1).txt
pandas
numpy
matplotlib
seaborn
jupyter
mysql.connector
squarify